requirements.txt
django
django-crispy-forms
django-agenda
djangorestframework
markdown
django-filter
django-agenda
django-bootstrap-datepicker-plus
django-intl-tel-input
